--story=1006222 --user=龙儒锋 profile所属系统添加全局引用、profile引用全局参数 https://www.tapd.cn/54247054/s/1228500

+++ src/main/java/codedriver/module/autoexec/api/profile/AutoexecProfileSaveApi.java
@@ -73,12 +73,12 @@ public class AutoexecProfileSaveApi extends PrivateApiComponentBase {
             throw new AutoexecProfileNameRepeatsException(profileVo.getName());
         }
 
-        //删除profile和tool、script的关系
-        autoexecProfileMapper.deleteProfileOperationByProfileId(paramProfileId);
-        //保存profile和tool、script的关系
-        autoexecProfileService.saveProfileOperation(profileVo.getId(), profileVo.getAutoexecOperationVoList());
-        autoexecProfileService.insertProfile(profileVo);
-//        autoexecProfileMapper.insertProfileParam();
+        if (paramProfileId != null) {
+            //删除profile引用tool、script的关系
+            autoexecProfileMapper.deleteProfileOperationByProfileId(paramProfileId);
+        }
+        //保存profile、profile参数、profile参数值引用全局参数的关系、profile引用tool、script的关系
+        autoexecProfileService.saveProfile(profileVo);
         return null;
     }
